Configure logger on message received
